feat: extract 34 inline routes from server/index.js into modular route files
Fixes #1240
