+++ server/routes/api.js
@@ -8,6 +8,7 @@ import * as usersController from '../controllers/usersController.js';
 import { adminAuditMiddleware, attachOldState } from '../middleware/adminAuditMiddleware.js';
 import { eventsRepository } from '../repositories/eventsRepository.js';
 import { coreTeamService } from '../services/coreTeamService.js';
+import { authRateLimiter, protectedActionRateLimiter } from '../middleware/authRateLimiter.js';
 
 const router = Router();
 
@@ -21,18 +22,20 @@ router.get(
 );
 router.post(
   '/api/content/activity-events/:activityKey',
+  protectedActionRateLimiter,
   adminAuthMiddleware.requireScope('events:write'),
   activityEventsController.addActivityEvent
 );
 router.delete(
   '/api/content/activity-events/:activityKey/:eventId',
+  protectedActionRateLimiter,
   adminAuthMiddleware.requireScope('events:write'),
   activityEventsController.deleteActivityEvent
 );
 
 // Admin auth
 router.get('/api/admin/users', adminAuthMiddleware.requireAdmin, usersController.getAdminUsers);
-router.post('/api/admin/login', adminAuthMiddleware.login);
+router.post('/api/admin/login', authRateLimiter, adminAuthMiddleware.login);
 router.post('/api/admin/logout', adminAuthMiddleware.requireAdmin, adminAuthMiddleware.logout);
 
 router.get(
